requirements.txt
discord.py>=2.2.2
requests
python-dotenv
ossapi